+++ cmd/config.go
@@ -17,9 +17,12 @@ This file is typically saved in your project root directory as golp.yaml
 All styles, scripts, and assets are relative to this file.`,
 	Args: cobra.ExactArgs(0),
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println(`# optional - list of folders or files to be deleted on every build / clean
+		fmt.Println(`# Optionally specify directories and/or files to automatically delete on every build / clean
 clean: 
   - themes/site/dist
+
+# Optionally run all copied images through image optimisers.
+optimise_images: true
 
 # SASS & CSS files
 styles: