+++ cmd/watch.go
@@ -17,6 +17,12 @@ This will monitor your src directories for changes and instantly rebuild their
 assets when a change is detected.`,
 	Args: cobra.ExactArgs(0),
 	Run: func(cmd *cobra.Command, args []string) {
+		if app.QuietLogging && app.VerboseLogging {
+			app.Log().Error("--quiet and --verbose are mutualfly exclusive\n\n")
+			_ = cmd.Help()
+			os.Exit(1)
+		}
+
 		if err := app.ParseConfig(); err != nil {
 			app.Log().Error(err.Error())
 			os.Exit(1)